+++ package.json
@@ -12,11 +12,11 @@
         "test": "npx qunit-testrunner",
         "test:keepAlive": "npx qunit-testrunner --keepalive",
         "test:cov": "npm run build:cov && npx qunit-testrunner --cov",
-        "test:dev": "NODE_ENV=development run-p test:keepAlive build:watch",
+        "test:dev": "cross-env NODE_ENV=development run-p test:keepAlive build:watch",
         "coverage": "nyc report",
         "coverage:html": "nyc report --reporter=lcov && open-cli coverage/lcov-report/index.html",
         "build": "rollup --config ./build/rollup.config.js",
-        "build:dev": "NODE_ENV=development npm run build:watch",
+        "build:dev": "cross-env NODE_ENV=development npm run build:watch",
         "build:cov": "rollup --config ./build/rollup.config.js --environment COVERAGE",
         "build:watch": "rollup --config ./build/rollup.config.js --watch",
         "build:scss": "node ./build/scss.js",
