+++ src/plugins/content/accessibility/jumplinks/jumplinks.tpl
@@ -1,0 +1,19 @@
+<nav class="jump-links-box" aria-label="{{__ "Jump Menu"}}">
+    <ul>
+        <li data-control="jump-link-question">
+            <a role="button" class="jump-link-question" href="#">{{__ "Jump to:"}} <b>{{__ "Question"}}</b></a>
+        </li>
+        <li data-control="jump-link-navigation">
+            <a role="button" class="jump-link-navigation" href="#">{{__ "Jump to:"}} <b>{{__ "Navigation"}}</b></a>
+        </li>
+        <li data-control="jump-link-toolbox">
+            <a role="button" class="jump-link-toolbox" href="#">{{__ "Jump to:"}} <b>{{__ "Toolbox"}}</b></a>
+        </li>
+        <li data-control="jump-link-teststatus">
+            <a role="button" class="jump-link-teststatus" href="#">{{__ "Jump to:"}} <b>{{__ "Test Status and Structure"}}</b></a>
+        </li>
+        <li data-control="jump-link-shortcuts">
+            <a role="button" class="jump-link-shortcuts" href="#">{{__ "Open Keyboard Shortcuts"}}</a>
+        </li>
+    </ul>
+</nav>
