+++ src/plugins/content/accessibility/jumplinks/shortcuts.scss
@@ -1,0 +1,75 @@
+/**
+ * Copyright (c) 2020 Open Assessment Technologies SA
+*/
+@import 'inc/bootstrap';
+
+.action-bar {
+
+  .jump-links-box {
+    display: block;
+    width: 0px;
+    height: 0px;
+    position: absolute;
+
+    li[data-control^=jump-link] {
+      position: fixed;
+      display: block;
+      z-index: 1001;
+    }
+
+    [class^=jump-link] {
+      z-index: 1002;
+      position: fixed;
+      display: block;
+      overflow: hidden;
+      top: 10px;
+      left: 10px;
+      width: 0px;
+      padding: 0px !important;
+      @include font-size(24);
+
+      &:focus {
+        width: auto;
+        height: auto;
+        padding: 30px !important;
+        background: $uiClickableDefaultBg;
+        color: $textColor !important;
+        border: 3px solid $textColor;
+        outline: none;
+        text-shadow: none !important;
+      }
+      &:focus:hover {
+        background: $uiClickableDefaultBg;
+        text-decoration: underline;
+      }
+    }
+  }
+
+  .shortcuts-list-wrapper {
+    width: 100%;
+    height: 100%;
+    position: fixed;
+    top: 0;
+    left: 0;
+    display: flex;
+    align-items: center;
+    align-content: center;
+    justify-content: center;
+    overflow: auto;
+    background: rgba(0,0,0,0.5);
+    z-index: 1002;
+
+    .shortcuts-list {
+      width: auto;
+      min-width: 400px;
+      height: auto;
+      min-height: 300px;
+      padding: 30px !important;
+      background: $uiClickableDefaultBg;
+      color: $textColor !important;
+      border: 3px solid $textColor;
+      text-shadow: none !important;
+    }
+  }
+
+}
