+++ src/plugins/content/accessibility/jumplinks/shortcuts.tpl
@@ -1,0 +1,85 @@
+<div class="shortcuts-list-wrapper hidden">
+    <div class="shortcuts-list">
+        <h1 id="keyboard-navigation">Keyboard Navigation</h1>
+        <blockquote>
+            <p>Keyboard shortcuts for the Accessibility Tools are available to the Test-taker.</p>
+        </blockquote>
+        <p>The set of keyboard shortcuts provided is as follows:</p>
+        <table>
+            <thead>
+            <tr>
+                <th><strong>Tool</strong></th>
+                <th>Action</th>
+                <th style="text-align: right;">Shortcut</th>
+            </tr>
+            </thead>
+            <tbody>
+            <tr>
+                <td><strong>Next Item</strong></td>
+                <td></td>
+                <td style="text-align: right;">J</td>
+            </tr>
+            <tr>
+                <td><strong>Previous Item</strong></td>
+                <td></td>
+                <td style="text-align: right;">K</td>
+            </tr>
+            <tr>
+                <td><strong>Answer Masking</strong></td>
+                <td>Toggle</td>
+                <td style="text-align: right;">D</td>
+            </tr>
+            <tr>
+                <td><strong>Area Masking</strong></td>
+                <td>Toggle</td>
+                <td style="text-align: right;">Y</td>
+            </tr>
+            <tr>
+                <td><strong>Calculator</strong></td>
+                <td>Toggle</td>
+                <td style="text-align: right;">C</td>
+            </tr>
+            <tr>
+                <td><strong>Highlighter</strong></td>
+                <td>Toggle</td>
+                <td style="text-align: right;">Shift + U</td>
+            </tr>
+            <tr>
+                <td><strong>Line Reader</strong></td>
+                <td>Toggle</td>
+                <td style="text-align: right;">G</td>
+            </tr>
+            <tr>
+                <td><strong>Magnifier</strong></td>
+                <td>Toggle</td>
+                <td style="text-align: right;">L</td>
+            </tr>
+            <tr>
+                <td><strong>Magnifier</strong></td>
+                <td>In</td>
+                <td style="text-align: right;">Shift + I</td>
+            </tr>
+            <tr>
+                <td><strong>Magnifier</strong></td>
+                <td>Out</td>
+                <td style="text-align: right;">Shift + O</td>
+            </tr>
+            <tr>
+                <td><strong>Magnifier</strong></td>
+                <td>Close</td>
+                <td style="text-align: right;">Esc</td>
+            </tr>
+            <tr>
+                <td><strong>Zoom</strong></td>
+                <td>In</td>
+                <td style="text-align: right;">I</td>
+            </tr>
+            <tr>
+                <td><strong>Zoom</strong></td>
+                <td>Out</td>
+                <td style="text-align: right;">O</td>
+            </tr>
+            </tbody>
+        </table>
+    </div>
+</div>
